+++ src/configuration/ConfigEditor.tsx
@@ -3,6 +3,8 @@ import { DataSourceHttpSettings, Input, InlineField, FieldSet } from '@grafana/u
 import { DataSourcePluginOptionsEditorProps, DataSourceSettings } from '@grafana/data';
 import { QuickwitOptions } from 'quickwit';
 import { coerceOptions } from './utils';
+import { Divider } from 'components/Divider';
+import { DataLinks } from './DataLinks';
 
 interface Props extends DataSourcePluginOptionsEditorProps<QuickwitOptions> {}
 
@@ -27,6 +29,7 @@ export const ConfigEditor = (props: Props) => {
         onChange={onOptionsChange}
       />
       <QuickwitDetails value={options} onChange={onSettingsChange} />
+      <QuickwitDataLinks value={options} onChange={onOptionsChange} />
     </>
   );
 };
@@ -35,6 +38,27 @@ type DetailsProps = {
   value: DataSourceSettings<QuickwitOptions>;
   onChange: (value: DataSourceSettings<QuickwitOptions>) => void;
 };
+
+export const QuickwitDataLinks = ({ value, onChange }: DetailsProps) => {
+  return (
+    <div className="gf-form-group">
+      <Divider hideLine />
+      <DataLinks
+        value={value.jsonData.dataLinks}
+        onChange={(newValue) => {
+          onChange({
+            ...value,
+            jsonData: {
+              ...value.jsonData,
+              dataLinks: newValue,
+            },
+          });
+        }}
+      />
+    </div>
+  )
+};
+
 export const QuickwitDetails = ({ value, onChange }: DetailsProps) => {
   return (
     <>
@@ -46,24 +70,6 @@ export const QuickwitDetails = ({ value, onChange }: DetailsProps) => {
               value={value.jsonData.index}
               onChange={(event) => onChange({ ...value, jsonData: {...value.jsonData, index: event.currentTarget.value}})}
               placeholder="otel-logs-v0"
-              width={40}
-            />
-          </InlineField>
-          <InlineField label="Timestamp field" labelWidth={26} tooltip="Timestamp field of your index. Required.">
-            <Input
-              id="quickwit_index_timestamp_field"
-              value={value.jsonData.timeField}
-              onChange={(event) => onChange({ ...value, jsonData: {...value.jsonData, timeField: event.currentTarget.value}})}
-              placeholder="timestamp"
-              width={40}
-            />
-          </InlineField>
-          <InlineField label="Timestamp output format" labelWidth={26} tooltip="If you don't remember the output format, check the datasource and the error message will give you a hint.">
-            <Input
-              id="quickwit_index_timestamp_field_output_format"
-              value={value.jsonData.timeOutputFormat}
-              onChange={(event) => onChange({ ...value, jsonData: {...value.jsonData, timeOutputFormat: event.currentTarget.value}})}
-              placeholder="unix_timestamp_millisecs"
               width={40}
             />
           </InlineField>